split amort renderer into classes
